Move TF2Filter to header and implementation files.,
